Optimization, cleanup, comments
This and the previous commit, switched the mod over to IPv6 with IPv4 backwards compatability.
Other improvements are having all ROUTE macros check for the address with `/` at the end by default now as well, unfortunately meaning that even /cover.jpg/ will give the cover image as it is now a valid URL.
Added some comments.
Multicast will send out IPv6 addresses for HTTP and the Socket on top of the IPv4 ones, look at the updated README.md for more information.
Improved DEBUG Logging defines.